requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
jwt==1.3.1
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
